requirements.txt
fastapi
uvicorn
jinja2
shortid-py
